# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "AddMemberBody": {
        "properties": {
          "dateOfBirth": {
            "format": "date",
            "type": "string"
          },
          "displayName": {
            "maxLength": 80,
            "minLength": 0,
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "relationship": {
            "type": "string"
          }
        },
        "required": [
          "displayName"
        ],
        "type": "object"
      },
...
      "CreateHouseholdBody": {
        "properties": {
          "defaultVisibility": {
            "type": "string"
          },
          "displayName": {
            "maxLength": 80,
            "minLength": 0,
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "HouseholdResponse": {
        "properties": {
          "baseCurrency": {
            "type": "string"
          },
          "defaultVisibility": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "memberCount": {
            "format": "int32",
            "type": "integer"
          },
          "myMemberId": {
            "format": "uuid",
            "type": "string"
          },
          "myRole": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "baseCurrency",
          "defaultVisibility",
          "id",
          "memberCount",
          "myRole",
          "name",
          "version"
        ],
        "type": "object"
      },
...
      "MemberResponse": {
        "properties": {
          "dateOfBirth": {
            "format": "date",
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "isManaged": {
            "type": "boolean"
          },
          "isMe": {
            "type": "boolean"
          },
          "isMinor": {
            "type": "boolean"
          },
          "relationship": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "displayName",
          "id",
          "isManaged",
          "isMe",
          "isMinor",
          "version"
        ],
        "type": "object"
      },
...
      "UpdateHouseholdBody": {
        "properties": {
          "defaultVisibility": {
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "UpdateMemberBody": {
        "properties": {
          "dateOfBirth": {
            "format": "date",
            "type": "string"
          },
          "displayName": {
            "maxLength": 80,
            "minLength": 0,
            "type": "string"
          },
          "relationship": {
            "type": "string"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "Access token from POST /api/v1/auth/otp/verify.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "padmini.koduri@bhrigu.tech",
      "name": "Almira"
    },
    "description": "Every rupee a family has invested or owes, in one private place.\n\n**Authentication.** Phone OTP issues a short-lived access token and a\nrotating refresh token. Send `Authorization: Bearer <accessToken>`.\nRefresh tokens are single-use: presenting one twice is treated as theft\nand ends the session.\n\n**Privacy.** Every read passes a per-record visibility check enforced by\nPostgreSQL row-level security. A record you may not see returns 404, not\n403 — a 403 would confirm it exists. Totals are computed through the same\nfilter, so a private record contributes nothing, not even its amount, to\nanother member's figures.\n\n**Money** is decimal, never floating point. Amounts also arrive formatted\n(`₹1,76,875`) so every surface renders them identically.\n\n**Concurrency.** Writes carry the record's `version`; a stale write is\nrejected with 409 and the current version. Creates accept a client-supplied\n`id`, so an offline capture keeps its identity and a retry is idempotent.\n\n**Stability.** v1 is frozen and additive-only. Breaking changes go to v2.",
    "title": "Almira API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/households": {
      "get": {
        "operationId": "listMine",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HouseholdResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "household-controller"
        ]
      },
      "post": {
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHouseholdBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdResponse"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "household-controller"
        ]
      }
    },
    "/api/v1/households/{householdId}": {
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "household-controller"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHouseholdBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "household-controller"
        ]
      }
    },
...
    "/api/v1/households/{householdId}/members": {
      "get": {
        "operationId": "members",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MemberResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "household-controller"
        ]
      },
      "post": {
        "operationId": "addMember",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "household-controller"
        ]
      }
    },
    "/api/v1/households/{householdId}/members/{memberId}": {
      "delete": {
        "operationId": "removeMember",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "household-controller"
        ]
      },
      "patch": {
        "operationId": "updateMember",
        "parameters": [
          {
            "in": "path",
            "name": "householdId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "household-controller"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ]
}